refactor(examples): move data paths out of initializer
