--- examples/taxonomy-classification/taxonomy_classification/__init__.py
@@ -1,1 +1,0 @@
-"""Taxonomy classification example package."""
