--- examples/taxonomy-classification/taxonomy_classification/classifier/__init__.py
@@ -1,1 +1,0 @@
-"""Classifier implementations for the taxonomy classification example."""
