--- examples/taxonomy-classification/taxonomy_classification/data/__init__.py
@@ -1,1 +1,0 @@
-"""Data preparation utilities for the taxonomy classification example."""
